langchain/requirements.txt
streamlit
audio-recorder-streamlit
openai-whisper
requests
python-dotenv
streamlit-chat
elevenlabs
